[lldb-dap] group android configuration builders
